+++ app/src/Security/LoginFormAuthenticator.php
@@ -1,0 +1,134 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Security;
+
+use App\Entity\User;
+use App\Repository\UserRepository;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
+use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+use Symfony\Component\Security\Core\Exception\CustomUserMessageAuthenticationException;
+use Symfony\Component\Security\Core\Exception\InvalidCsrfTokenException;
+use Symfony\Component\Security\Core\Security;
+use Symfony\Component\Security\Core\User\UserInterface;
+use Symfony\Component\Security\Core\User\UserProviderInterface;
+use Symfony\Component\Security\Csrf\CsrfToken;
+use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
+use Symfony\Component\Security\Guard\Authenticator\AbstractFormLoginAuthenticator;
+use Symfony\Component\Security\Guard\PasswordAuthenticatedInterface;
+use Symfony\Component\Security\Http\Util\TargetPathTrait;
+
+class LoginFormAuthenticator extends AbstractFormLoginAuthenticator implements
+    PasswordAuthenticatedInterface
+{
+    use TargetPathTrait;
+
+    private UserRepository $userRepo;
+    private UrlGeneratorInterface $urlGenerator;
+    private CsrfTokenManagerInterface $csrfTokenManager;
+    private UserPasswordEncoderInterface $passwordEncoder;
+
+    public function __construct(
+        UserRepository $userRepo,
+        UrlGeneratorInterface $urlGenerator,
+        CsrfTokenManagerInterface $csrfTokenManager,
+        UserPasswordEncoderInterface $passwordEncoder
+    ) {
+        $this->userRepo = $userRepo;
+        $this->urlGenerator = $urlGenerator;
+        $this->csrfTokenManager = $csrfTokenManager;
+        $this->passwordEncoder = $passwordEncoder;
+    }
+
+    public function supports(Request $request): bool
+    {
+        return $request->attributes->get('_route') === 'app_login'
+            && $request->isMethod('POST');
+    }
+
+    /** @return array<string> */
+    public function getCredentials(Request $request): array
+    {
+        $credentials = [
+            'email' => $request->request->get('email'),
+            'password' => $request->request->get('password'),
+            'csrf_token' => $request->request->get('_csrf_token'),
+        ];
+        $request->getSession()->set(
+            Security::LAST_USERNAME,
+            $credentials['email'],
+        );
+
+        return $credentials;
+    }
+
+    /**
+     * @param array<string> $credentials
+     * @phpcsSuppress SlevomatCodingStandard.TypeHints.ParameterTypeHint.MissingNativeTypeHint
+     * @phpcsSuppress SlevomatCodingStandard.Functions.UnusedParameter.UnusedParameter
+     */
+    public function getUser($credentials, UserProviderInterface $userProvider): User
+    {
+        $token = new CsrfToken('authenticate', $credentials['csrf_token']);
+        if (!$this->csrfTokenManager->isTokenValid($token)) {
+            throw new InvalidCsrfTokenException('CSRF token is invalid.');
+        }
+
+        $user = $this->userRepo->findOneBy(['email' => $credentials['email']]);
+
+        if (!$user instanceof User) {
+            throw new CustomUserMessageAuthenticationException(
+                'Email could not be found.',
+            );
+        }
+
+        return $user;
+    }
+
+    /**
+     * @param array<string> $credentials
+     * @phpcsSuppress SlevomatCodingStandard.TypeHints.ParameterTypeHint.MissingNativeTypeHint
+     */
+    public function checkCredentials($credentials, UserInterface $user): bool
+    {
+        $password = $credentials['password'];
+
+        return $this->passwordEncoder->isPasswordValid($user, $password);
+    }
+
+    /**
+     * Used to upgrade (rehash) the user's password automatically over time.
+     *
+     * @param array<string> $credentials
+     * @phpcsSuppress SlevomatCodingStandard.TypeHints.ParameterTypeHint.MissingNativeTypeHint
+     */
+    public function getPassword($credentials): ?string
+    {
+        return $credentials['password'];
+    }
+
+
+    /** @phpcsSuppress SlevomatCodingStandard.Functions.UnusedParameter.UnusedParameter */
+    public function onAuthenticationSuccess(
+        Request $request,
+        TokenInterface $token,
+        string $providerKey
+    ): Response {
+        $targetPath = $this->getTargetPath($request->getSession(), $providerKey);
+        if (is_string($targetPath)) {
+            return new RedirectResponse($targetPath);
+        }
+
+        return new RedirectResponse($this->urlGenerator->generate('test'));
+    }
+
+    protected function getLoginUrl(): string
+    {
+        return $this->urlGenerator->generate('app_login');
+    }
+}